Making fixes trying to run the linear quasigeostrophic code example.

+++ GLNumericalModelingKit/GLNetCDFFile.m
@@ -590,7 +590,7 @@ const static NSString *MutableDimensionContext = @"com.EarlyInnovations.MutableD
 	[self.dimensionVariableIDMapTable setObject: [NSNumber numberWithInteger: variableID] forKey: dimension];
 	NSArray *array = [NSArray arrayWithObject: [NSValue valueWithRange: NSMakeRange(0, dimension.nPoints)]];
 	if (sizeof(GLFloat)==sizeof(float)) {
-		[self.file writeFloatData: dimension.data toVariableWithID: variableID atIndexRange: array];
+		[self.file writeFloatData: dimension.data toVariableWithID: (int) variableID atIndexRange: array];
 	} else {
 		[self.file writeDoubleData: dimension.data toVariableWithID: variableID atIndexRange: array];
 	}
